+++ src/main/java/de/dm/intellij/liferay/language/freemarker/staticfieldgetter/StaticFieldGetterClassNameCompletionContributor.java
@@ -27,30 +27,30 @@ public class StaticFieldGetterClassNameCompletionContributor extends CompletionC
 
     public StaticFieldGetterClassNameCompletionContributor() {
         extend(
-                CompletionType.BASIC,
-                ELEMENT_FILTER,
-                new CompletionProvider<CompletionParameters>() {
-                    @Override
-                    protected void addCompletions(@NotNull CompletionParameters parameters, ProcessingContext context, @NotNull CompletionResultSet result) {
-                        PsiElement originalPosition = parameters.getOriginalPosition();
-                        if (originalPosition != null) {
-                            PsiFile psiFile = originalPosition.getContainingFile();
-                            psiFile = psiFile.getOriginalFile();
+            CompletionType.BASIC,
+            ELEMENT_FILTER,
+            new CompletionProvider<CompletionParameters>() {
+                @Override
+                protected void addCompletions(@NotNull CompletionParameters parameters, ProcessingContext context, @NotNull CompletionResultSet result) {
+                    PsiElement originalPosition = parameters.getOriginalPosition();
+                    if (originalPosition != null) {
+                        PsiFile psiFile = originalPosition.getContainingFile();
+                        psiFile = psiFile.getOriginalFile();
 
-                            Module module = ModuleUtil.findModuleForFile(psiFile);
+                        Module module = ModuleUtil.findModuleForFile(psiFile);
 
-                            if (isStaticFieldGetterCall(originalPosition)) {
-                                //TODO filter by classes having static methods
-                                PsiClass objectClass = ProjectUtils.getClassByName(originalPosition.getProject(), "java.lang.Object", originalPosition);
+                        if (isStaticFieldGetterCall(originalPosition)) {
+                            //TODO filter by classes having static methods
+                            PsiClass objectClass = ProjectUtils.getClassByName(originalPosition.getProject(), "java.lang.Object", originalPosition);
 
-                                LiferayFreemarkerUtil.addClassInheritorsLookup(objectClass, result, module);
+                            LiferayFreemarkerUtil.addClassInheritorsLookup(objectClass, result, module);
 
-                                result.stopHere();
-                            }
-
+                            result.stopHere();
                         }
+
                     }
                 }
+            }
         );
 
     }