+++ testdata/de/dm/intellij/liferay/language/freemarker/LiferayFtlVariableProviderTest/com/liferay/portal/kernel/util/PortletKeys.java
@@ -1,0 +1,8 @@
+package com.liferay.portal.kernel.util;
+
+public class PortletKeys {
+
+    public static final String ADMIN_PLUGINS = "136";
+    public static final String BACKGROUND_TASK = "189";
+
+}