+++ testdata/de/dm/intellij/liferay/language/freemarker/LiferayFtlVariableProviderTest/com/liferay/portal/kernel/util/StaticFieldGetter.java
@@ -1,0 +1,9 @@
+package com.liferay.portal.kernel.util;
+
+public class StaticFieldGetter {
+
+    public Object getFieldValue(String className, String fieldName) {
+        return null;
+    }
+
+}